Applies template pattern to reduce duplicate code in smartmetering requestMessageProcessors. Removes commented out code.
