+++ osgp-adapter-domain-smartmetering/src/main/java/com/alliander/osgp/adapter/domain/smartmetering/infra/jms/ws/messageprocessors/SetActivityCalendarRequestMessageProcessor.java
@@ -7,11 +7,6 @@
  */
 package com.alliander.osgp.adapter.domain.smartmetering.infra.jms.ws.messageprocessors;
 
-import javax.jms.JMSException;
-import javax.jms.ObjectMessage;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
@@ -20,12 +15,10 @@ import com.alliander.osgp.adapter.domain.smartmetering.application.services.Conf
 import com.alliander.osgp.adapter.domain.smartmetering.infra.jms.ws.WebServiceRequestMessageProcessor;
 import com.alliander.osgp.domain.core.valueobjects.DeviceFunction;
 import com.alliander.osgp.domain.core.valueobjects.smartmetering.ActivityCalendar;
-import com.alliander.osgp.shared.infra.jms.Constants;
+import com.alliander.osgp.shared.exceptionhandling.FunctionalException;
 
 @Component("domainSmartmeteringSetActivityCalendarRequestMessageProcessor")
 public class SetActivityCalendarRequestMessageProcessor extends WebServiceRequestMessageProcessor {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(SetActivityCalendarRequestMessageProcessor.class);
 
     @Autowired
     @Qualifier("domainSmartMeteringConfigurationService")
@@ -35,49 +28,13 @@ public class SetActivityCalendarRequestMessageProcessor extends WebServiceReques
         super(DeviceFunction.SET_ACTIVITY_CALENDAR);
     }
 
-    /*
-     * (non-Javadoc)
-     *
-     * @see
-     * com.alliander.osgp.shared.infra.jms.MessageProcessor#processMessage(javax
-     * .jms.ObjectMessage)
-     */
     @Override
-    public void processMessage(final ObjectMessage message) throws JMSException {
+    protected void handleMessage(final String organisationIdentification, final String deviceIdentification,
+            final String correlationUid, final Object dataObject, final String messageType) throws FunctionalException {
 
-        String correlationUid = null;
-        String messageType = null;
-        String organisationIdentification = null;
-        String deviceIdentification = null;
-        Object dataObject = null;
+        final ActivityCalendar activityCalendar = (ActivityCalendar) dataObject;
 
-        try {
-            correlationUid = message.getJMSCorrelationID();
-            messageType = message.getJMSType();
-            organisationIdentification = message.getStringProperty(Constants.ORGANISATION_IDENTIFICATION);
-            deviceIdentification = message.getStringProperty(Constants.DEVICE_IDENTIFICATION);
-            dataObject = message.getObject();
-
-        } catch (final JMSException e) {
-            LOGGER.error("UNRECOVERABLE ERROR, unable to read ObjectMessage instance, giving up.", e);
-            LOGGER.debug("correlationUid: {}", correlationUid);
-            LOGGER.debug("messageType: {}", messageType);
-            LOGGER.debug("organisationIdentification: {}", organisationIdentification);
-            LOGGER.debug("deviceIdentification: {}", deviceIdentification);
-            return;
-        }
-
-        try {
-            LOGGER.info("Calling application service function: {}", messageType);
-
-            final ActivityCalendar activityCalendar = (ActivityCalendar) dataObject;
-
-            this.configurationService.setActivityCalendar(organisationIdentification, deviceIdentification,
-                    correlationUid, activityCalendar, messageType);
-
-        } catch (final Exception e) {
-            this.handleError(e, correlationUid, organisationIdentification, deviceIdentification, messageType);
-        }
-
+        this.configurationService.setActivityCalendar(organisationIdentification, deviceIdentification, correlationUid,
+                activityCalendar, messageType);
     }
 }
